Make prop names, values and component texts (halfway) editable
For prop names, we're simply rendering a clickable minus symbol like
before. For values, we're rendering an on-click popup with a
property-type-specific value chooser. For component texts we are
temporarily rendering a clickable minus symbol, although this is
mainly as a proof of concept.

All actions result in a new component tree that is passed back up
to the ComponentTreeEditor via the plugin's onChange function prop.